# internal/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Skifity API
  version: development
  summary: Self-hosted apps, powered by Kubernetes.
  description: |
    The API the panel, the `skifity` CLI and the MCP server all use. Anything the
    panel can do, a request here can do.

    **Authentication.** Send an API token as `Authorization: Bearer skf_...`.
    Create one in the panel under Account, then API tokens. A browser session
    works too; it is the `skifity_session` cookie (`__Host-skifity_session` on
    HTTPS), and a request that changes something then also carries the
    session's CSRF token in the `X-Skifity-CSRF` header. A token needs no CSRF
    header. A few actions refuse a token and want a person who has just typed
    their password; those list only the session scheme.

    **Answers.** Lists come back as `{"items": [...], "total": n}`. Every
    failure has the same shape: an `error` object with a stable `code`, a
    `title`, and the `cause`, `impact` and `fix` in words to act on.

    **Streams.** The event stream, and an app's logs followed live, are
    server-sent events (`text/event-stream`).

    **Roles.** A team member is an owner, an admin, a member or a viewer. A
    viewer reads everything a member can and changes nothing. Routes marked as
    for panel administrators want an account that administers the whole
    panel, whatever its role in a team.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
servers:
  - url: /
...

...
security:
  - bearerToken: []
  - sessionCookie: []
...
paths:
...
  /api/teams/{teamID}/secret-managers/{connectionID}/test:
    parameters:
      - $ref: '#/components/parameters/teamID'
      - $ref: '#/components/parameters/connectionID'
    post:
      operationId: testSecretManager
      tags: [Secret managers]
      summary: Check that a connection still signs in
      description: >-
        Admin. Vault's token lookup, Infisical's login, Doppler's token
        details, or AWS's GetCallerIdentity: nothing is read.
      responses:
        '200': {$ref: '#/components/responses/OK'}
        default: {$ref: '#/components/responses/Error'}

  # --- servers ---
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: skf_...
      description: >-
        An API token, made in the panel under Account, then API tokens. It is
        made for one team, can do no more than its owner's role in that team,
        and may be narrowed further with scopes. It needs no CSRF header.
    sessionCookie:
      type: apiKey
      in: cookie
      name: skifity_session
      description: >-
        A browser session, set by signing in. On HTTPS the cookie is
        `__Host-skifity_session`. A request that changes something also sends
        the session's CSRF token, which signing in returns as csrf_token, in
        the X-Skifity-CSRF header.
  parameters:
    teamID: {name: teamID, in: path, required: true, description: 'A team''s id.', schema: {type: string}}
...
    connectionID: {name: connectionID, in: path, required: true, description: 'A secret manager connection''s id, or its name.', schema: {type: string}}
...
  responses:
    Error:
      description: >-
        What went wrong, what it means and what to do. The code is stable; the
        sentences are English, and args carries the values in them so the
        panel can say the same thing in another language.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    OK:
      description: Done. There is nothing else to say.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/OK'}
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: {$ref: '#/components/schemas/Problem'}
    Problem:
      type: object
      description: A failure with its explanation. See internal/errdoc.
      required: [code, title, severity, retryable, at]
      properties:
        code: {type: string, description: 'Stable and machine-readable, such as deploy.build_failed.'}
        title: {type: string}
        cause: {type: string, description: 'What actually went wrong.'}
        impact: {type: string, description: 'What is and is not working because of it.'}
        fix: {type: string, description: 'What to do about it.'}
        args:
          type: object
          description: The values in each sentence, in order, for a translation to put in the same places.
          properties:
            title: {type: array, items: {type: string}}
            cause: {type: array, items: {type: string}}
            impact: {type: array, items: {type: string}}
            fix: {type: array, items: {type: string}}
        docs_path: {type: string, description: 'A page of the documentation the panel serves, such as /docs/troubleshooting#a-deployment-failed.'}
        severity: {type: string, enum: [error, warning, info]}
        retryable: {type: boolean, description: 'Whether trying again might work.'}
        context:
          type: object
          description: Structured detail, such as the host or the exit code. Never a secret.
          additionalProperties: {type: string}
        at: {type: string, format: date-time}
    OK:
      type: object
      required: [ok]
      properties:
        ok: {type: boolean, const: true}
...